Moved ValueSetPicker to Grasshopper.External.Special namespace.
