+++ src/RhinoInside.Revit.GH/GH/ValueList.cs
@@ -5,166 +5,27 @@ using System.Linq;
 using System.Windows.Forms;
 using Grasshopper.GUI;
 using Grasshopper.GUI.Canvas;
-using Grasshopper.Kernel;
-using Grasshopper.Kernel.Attributes;
-using Grasshopper.Kernel.Data;
-using Grasshopper.Kernel.Special;
-using Grasshopper.Kernel.Types;
 
-namespace RhinoInside.Revit.GH.Parameters
+///////////////////////////////////////////////////////////////////////////////
+//                                                                           //
+// NOTE: At some point this code may end up in the Grasshopper code base.    //
+//                                                                           //
+///////////////////////////////////////////////////////////////////////////////
+
+namespace Grasshopper.External.Special
 {
-  public abstract class GH_ValueList : Grasshopper.Kernel.Special.GH_ValueList
-  {
-    protected override Bitmap Icon => ((Bitmap) Properties.Resources.ResourceManager.GetObject(GetType().Name)) ??
-                                       ImageBuilder.BuildIcon(GetType().Name.Substring(0, 1));
-  }
+  using Kernel;
+  using Kernel.Attributes;
+  using Kernel.Data;
+  using Kernel.Types;
 
-  public abstract class ValueList : GH_ValueList, IGH_InitCodeAware
-  {
-    public override GH_ParamData DataType => GH_ParamData.remote;
-    public override GH_ParamKind Kind => GH_ParamKind.floating;
-
-    void IGH_InitCodeAware.SetInitCode(string code) => NickName = code;
-
-    public ValueList()
-    {
-      Category = "Revit";
-      SubCategory = "Input";
-
-      ObjectChanged += OnObjectChanged;
-
-      ListMode = GH_ValueListMode.DropDown;
-      ListItems.Clear();
-    }
-
-    private void OnObjectChanged(IGH_DocumentObject sender, GH_ObjectChangedEventArgs e)
-    {
-      if (SourceCount == 0)
-      {
-        if (e.Type == GH_ObjectEventType.Sources)
-        {
-          //NickName = string.Empty;
-          MutableNickName = true;
-        }
-
-        if (e.Type == GH_ObjectEventType.NickName)
-          ExpireSolution(true);
-      }
-      else
-      {
-        if (e.Type == GH_ObjectEventType.Sources)
-        {
-          NickName = string.Empty;
-          MutableNickName = false;
-        }
-      }
-    }
-
-    class ValueListAttributes : GH_ValueListAttributes
-    {
-      public override bool HasInputGrip => true;
-      public override bool AllowMessageBalloon => true;
-      public ValueListAttributes(ValueList owner) : base(owner) { }
-      protected override void Render(GH_Canvas canvas, Graphics graphics, GH_CanvasChannel channel)
-      {
-        if (channel == GH_CanvasChannel.Wires)
-        {
-          if (Owner.SourceCount > 0)
-            RenderIncomingWires(canvas.Painter, Owner.Sources, Owner.WireDisplay);
-        }
-
-        base.Render(canvas, graphics, channel);
-      }
-      public override GH_ObjectResponse RespondToMouseDoubleClick(GH_Canvas sender, GH_CanvasMouseEvent e)
-      {
-        if (Owner.MutableNickName && e.Button == MouseButtons.Left)
-        {
-          var objectMenu = new ContextMenuStrip();
-
-          Owner.AppendMenuItems(objectMenu);
-          if (objectMenu.Items.Count > 0)
-          {
-            sender.ActiveInteraction = null;
-            objectMenu.Show(sender, e.ControlLocation);
-          }
-
-          return GH_ObjectResponse.Handled;
-        }
-
-        return GH_ObjectResponse.Ignore;
-      }
-    }
-
-    public override void CreateAttributes() => m_attributes = new ValueListAttributes(this);
-
-    protected abstract void RefreshList(string FamilyName);
-
-    protected abstract void RefreshList(IEnumerable<IGH_Goo> goos);
-
-    public override void PostProcessData()
-    {
-      base.PostProcessData();
-
-      if (SourceCount == 0)
-        RefreshList(NickName);
-      else
-        RefreshList(VolatileData.AllData(true));
-
-      // Show elements sorted
-      ListItems.Sort((x, y) => string.CompareOrdinal(x.Name, y.Name));
-
-      //base.CollectVolatileData_Custom();
-      m_data.Clear();
-
-      var path = new GH_Path(0);
-      if (SelectedItems.Count == 0)
-        m_data.AppendRange(new IGH_Goo[0], path);
-      else foreach (var item in SelectedItems)
-          m_data.Append(item.Value, path);
-    }
-
-    protected override void CollectVolatileData_FromSources()
-    {
-      base.CollectVolatileData_FromSources();
-
-      NickName = string.Empty;
-    }
-
-    protected override void CollectVolatileData_Custom()
-    {
-      NickName = NickName.Trim();
-    }
-
-    protected override string HtmlHelp_Source()
-    {
-      var nTopic = new Grasshopper.GUI.HTML.GH_HtmlFormatter(this)
-      {
-        Title = Name,
-        Description =
-        @"<p>Double click on it and use the name input box to enter a name, alternativelly you can enter a name patter. " +
-        @"If a pattern is used, this param list will be filled up with all the objects that match it.</p>" +
-        @"<p>Several kind of patterns are supported, the method used depends on the first pattern character:</p>" +
-        @"<dl>" +
-        @"<dt><b>></b></dt><dd>Starts with</dd>" +
-        @"<dt><b><</b></dt><dd>Ends with</dd>" +
-        @"<dt><b>?</b></dt><dd>Contains, same as a regular search</dd>" +
-        @"<dt><b>:</b></dt><dd>Wildcards, see Microsoft.VisualBasic " + "<a target=\"_blank\" href=\"https://docs.microsoft.com/en-us/dotnet/visual-basic/language-reference/operators/like-operator#pattern-options\">LikeOperator</a></dd>" +
-        @"<dt><b>;</b></dt><dd>Regular expresion, see " + "<a target=\"_blank\" href=\"https://docs.microsoft.com/en-us/dotnet/standard/base-types/regular-expression-language-quick-reference\">here</a> as reference</dd>" +
-        @"</dl>",
-        ContactURI = @"https://discourse.mcneel.com/c/serengeti/inside"
-      };
-
-      return nTopic.HtmlFormat();
-    }
-  }
-
-  public abstract class ValueSet : Grasshopper.Kernel.GH_PersistentParam<IGH_Goo>, IGH_InitCodeAware
+  public abstract class ValueSet : GH_PersistentParam<IGH_Goo>, IGH_InitCodeAware
   {
     public override string TypeName => "Data";
     public override GH_Exposure Exposure => GH_Exposure.secondary;
     public override GH_ParamKind Kind => GH_ParamKind.floating;
     protected override Bitmap Icon => ClassIcon;
-    static Bitmap ClassIcon => ImageBuilder.BuildIcon
+    static Bitmap ClassIcon => RhinoInside.Revit.ImageBuilder.BuildIcon
     (
       (graphics, bounds) =>
       {
@@ -660,7 +521,7 @@ namespace RhinoInside.Revit.GH.Parameters
 
       foreach (var goo in PersistentData)
       {
-        if (goo is Types.IGH_ElementId id) id.UnloadElement();
+        if (goo is RhinoInside.Revit.GH.Types.IGH_ElementId id) id.UnloadElement();
         else if (goo is IGH_GeometricGoo geo) geo.ClearCaches();
       }
     }
@@ -700,7 +561,7 @@ namespace RhinoInside.Revit.GH.Parameters
         {
           var goo = branch[i];
 
-          if (goo is Types.IGH_ElementId id && id.IsReferencedElement && !id.IsElementLoaded && !id.LoadElement())
+          if (goo is RhinoInside.Revit.GH.Types.IGH_ElementId id && id.IsReferencedElement && !id.IsElementLoaded && !id.LoadElement())
           {
             AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, $"A referenced element could not be found in the Revit document.");
             branch[i] = null;
@@ -883,14 +744,14 @@ namespace RhinoInside.Revit.GH.Parameters
       if (DataType == GH_ParamData.local)
       {
         ListItems = goosSet.Select(goo => new ListItem(goo, true)).
-                    Where(x => string.IsNullOrEmpty(NickName) || x.Name.IsSymbolNameLike(NickName)).
+                    Where(x => string.IsNullOrEmpty(NickName) || RhinoInside.Revit.Operator.IsSymbolNameLike(x.Name, NickName)).
                     ToList();
       }
       else if (DataType == GH_ParamData.remote)
       {
         var selectSet = new HashSet<IGH_Goo>(PersistentData.Where(x => GooComparer.IsComparable(x)), new GooComparer());
         ListItems = goosSet.Select(goo => new ListItem(goo, selectSet.Contains(goo))).
-                    Where(x => string.IsNullOrEmpty(NickName) || x.Name.IsSymbolNameLike(NickName)).
+                    Where(x => string.IsNullOrEmpty(NickName) || RhinoInside.Revit.Operator.IsSymbolNameLike(x.Name, NickName)).
                     ToList();
       }
       else
